Move connection, send and receive strategies out of SocketNetwork
